Add LoopRotate insertBackedgeBlock.
Run a loopSimplify helper before loopRotate to put the loop in
canonical form. This now handles the case of multiple-backedges, which
is very common for unsimplified loops.

TODO: We still don't split nested loops that have a shared
header. LLVM has logic for this, we should probably do the same.

Swift SVN r24008