Directly include ErrorOr.h and MemoryBuffer.h to make Passes.cpp resilient against potential forward declarations.
In the swift-602-branch, these are forward declared in one of the
headers included in this file yielded a compilation error. Rather than
just fixing the compilation failure on that branch, this commit makes
the entire file more redundent since who knows in the future these two
types may be forward declared again.

Swift SVN r24066