Improve the inlining heuristic for cases where a closure is passed to a function.
This change adds a general method to see if inlining would enable constant propagation or
inlining of a closure.
For this it does not matter if a constant/function_ref/etc. is passed directly or within
a struct.
This first version only handles closures which are passed to an apply in the callee.

It fixes the performance problem of RangeAssignment (rdar://problem/19252374) and shows
minor improvements in some other benchmarks, e.g. CommonMarkRender.

The impact on code size is negligible (< 1%).



Swift SVN r24109