Specialize the memcpy in swift_getEnumCaseSinglePayload for the case of small constant values.
The libc memcpy implementation is not optimized for the case of constant count
values (because of the function's interface) and is not inlineable. Using the local
implementation allows us to eliminate the trampoline in the call and optimize for
small constants.

This gives a 4% boost on Prims.

Swift SVN r24131