[ls-opts] Fix crasher where we are determing whether or not to load store forward while performing the load store forwarding. This can cause us to dereference an empty SILValue.
The forwarding feasability code in LoadStoreOpts is separated into an
analysis and a action part. The action part assumes that SILValues
returned are never null. This violates a normal error handling pattern
with SILValues. We should introduce some notion of a SILValue that can
not be null. Until then, this patch fixes the crasher.

<rdar://problem/19369189>

Swift SVN r24169