[ls-opt] Add in LSValue, LSStore, LSLoad and use it when we are not trying to forward from multiple reaching stores.
LSValue, LSStore, and LSLoad are structs that are meant to be able to
allow for load store opts to use the same logic for performing load
store opts from one load, store and covering sets of loads, stores.

This changes LoadStoreOpts to use these data structures when it is not
considering covering stores.

Swift SVN r24183