Bridged specialized uses of Objective-C generic collections to more-specialized Swift collections.
When importing a specialized type for an Objective-C collection (e.g.,
NSArray<NSString *> *) that is being bridged to a Swift collection,
produce a more specialized Swift collection type (e.g., [String]) that
we would for the unspecialized type ([NSObject]).

The CMake and lit hackery is here because we need to be able to build
against versions of Clang that both do and do not have Objective-C
generics.

Swift SVN r24196