pipeline: Don't hard code pass names into pipelines.py. Just use the PASSES variable from passes.py in the library.
The pipeline code still has too much hard coded. I am going to write a little
tool that is disabled by default that when run provides all this information
from the compiler. So the tool will dump out the current pass list, the normal optimization
pipelines, and then all of the passes used in such pipelines (i.e. AADumper is a
great pass but it won't change how the stdlib compiles). But for now I am going
to leave it hard coded.

Swift SVN r24313