Rename normal_pipeline.py => pipeline_generator.py and pipelines.py => pipelines_build_script.py.
This makes it clearer via their name what the two scripts are meant to
be used for.

As a recap:

1. pipeline_generator.py is meant for generating pass pipeline json
files.
2. pipelines_build_script.py is a script built ontop of
pipeline_generator.py that makes it easy to run build-script with
various pass pipelines that can be generated from
pipeline_generator.py. It is meant to be used on buildbots and for ones
own enjoyment as well.

Swift SVN r24315