Add the ability for a owning addressor to return a non-native owner.  This is required by Slice, which will use an ObjC immutable array object as the owner as long as all the elements are contiguous.
As part of this, I decided it was best to encode the
native requirement in the accessor names.  This makes
some of these accessors really long; we can revisit this
if we productize this feature.

Note that pinning addressors still require a native
owner, since pinning as a feature is specific to swift
refcounting.

Swift SVN r24420