SILGen: Initializing constructors retain the +1 calling convention for self.
The 'self' parameter will be forwarded back to the allocating entry point at +1, and we shouldn't have to retain it to do so. Furthermore, @objc constructors may replace "self", in which case they need to be able to deallocate the original "self". For these reasons, initializing constructors should still receive self at +1 even when guaranteed self is enabled.

Swift SVN r24430