[stdlib] fixed converting integers to floats at bounds
We found converting ints to floats at bounds caused it to overflow
without trapping. This was due to the construction of floats using
rounding (on x86 using an instruction dependent on the status register).

To fix this we check the bounds of the float with constants that are
known to be the last valid floating point numbers valid for conversion.

Fixes rdar://problem/18406402

Swift SVN r24459