Allow the right-hand side of same-type constraints to be an arbitrary type.
Allows same-type constraints that map down to tuple types. This is a
minimal fix that allows the code in rdar://problem/18120419 to
type-check. However, the actual code in that radar triggers a SILGen
assertion (Archetype to be substituted must be valid in function), and
we're not recursively matching concrete types the way we should be, so
this is a baby step.

Among other things, this allows one to provide a same-type con

Swift SVN r24535