IRGen: Form OpaqueStorage using [N x i8] arrays for more accurate sizing.
If the alloc size of a single iNNN field doesn't match the requested size, then it will use the wrong amount of space inside an LLVM struct. Solve this by using [N x i8] as the storage type for opaque storage, but loading and storing as a single iNNN field.

Swift SVN r24571