[docs] Start describing a hypothetical persistable module format.
The current serialization format has bits of groundwork for this, but
not a coherent implementation that's really close to being future-proof
and forward-compatible. We will eventually need this as part of being
able to ship compiled frameworks to people.

It is worth the time to get this out of my head and into a proper document.
I'll continue adding information later.

Swift SVN r24572