Work around a miscompile in NominalTypeDescriptor.FieldNames on arm64.
If the FieldNames data is exactly 4 or 8 or 16 bytes long then on arm64 
it is compiled as a 4/8/16 byte literal instead of as ordinary data. 
This confuses something in the toolchain so the final linked executable 
gets an invalid FieldNames pointer. 
Workaround: pad the FieldNames data to avoid those sizes.

rdar://19537198


Swift SVN r24587