SILCombine: Peephole a pattern generate by materializeForSetCallback for generic struct types.
This is not as general as it should be. The general solution is to introduce
a builtin partial_apply_thin_recoverable as John suggested to me in person and
describes in SILGenBuiltin.cpp.

This patch addresses an important subset of problems (ie. sorting an arrays of
things) with low risk for the upcoming release.

radar://19388835

Swift SVN r24628