SILCombine: Peephole a pattern generate by materializeForSetCallback
This reapplies commit r24628 with a fix: use the metatype of a boundgenerictype
to get the substitutions instead of painfully building them ourselves and making
mistakes along the way.

This is not as general as it should be. The general solution is to introduce a
builtin partial_apply_thin_recoverable as John suggested to me in person and
describes in SILGenBuiltin.cpp.

This patch addresses an important subset of problems (ie. sorting an arrays of
things) with low risk for the upcoming release.

radar://19388835

Swift SVN r24657